// clients/js/src/generated/types/baseAssetLinkedSecureDataStore.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { Serializer, struct } from '@metaplex-foundation/umi/serializers';
import {
  BasePluginAuthority,
  BasePluginAuthorityArgs,
  ExternalPluginAdapterSchema,
  ExternalPluginAdapterSchemaArgs,
  getBasePluginAuthoritySerializer,
  getExternalPluginAdapterSchemaSerializer,
} from '.';

export type BaseAssetLinkedSecureDataStore = {
  dataAuthority: BasePluginAuthority;
  schema: ExternalPluginAdapterSchema;
};

export type BaseAssetLinkedSecureDataStoreArgs = {
  dataAuthority: BasePluginAuthorityArgs;
  schema: ExternalPluginAdapterSchemaArgs;
};

export function getBaseAssetLinkedSecureDataStoreSerializer(): Serializer<
  BaseAssetLinkedSecureDataStoreArgs,
  BaseAssetLinkedSecureDataStore
> {
  return struct<BaseAssetLinkedSecureDataStore>(
    [
      ['dataAuthority', getBasePluginAuthoritySerializer()],
      ['schema', getExternalPluginAdapterSchemaSerializer()],
    ],
    { description: 'BaseAssetLinkedSecureDataStore' }
  ) as Serializer<
    BaseAssetLinkedSecureDataStoreArgs,
    BaseAssetLinkedSecureDataStore
  >;
}